refactor(AuthController,AuthService,LoginRequest,LoginResponse,RegisterRequest,Playlist,PlaylistVideo,PlaylistVideoId,Role,User,UserRepository,Video): refactor folder structure
