--- src/main/java/com/FemCoders/DropTheLink/models/Role.java
@@ -1,5 +1,0 @@
-package com.FemCoders.DropTheLink.models;
-
-public enum Role {
-    USER
-}
