+++ src/main/java/com/FemCoders/DropTheLink/playlist/model/Playlist.java
@@ -1,5 +1,6 @@
-package com.FemCoders.DropTheLink.models;
+package com.FemCoders.DropTheLink.playlist.model;
 
+import com.FemCoders.DropTheLink.user.User;
 import jakarta.persistence.*;
 import lombok.*;
 import java.util.ArrayList;
