+++ src/main/java/com/FemCoders/DropTheLink/playlist/model/PlaylistVideo.java
@@ -1,5 +1,6 @@
-package com.FemCoders.DropTheLink.models;
+package com.FemCoders.DropTheLink.playlist.model;
 
+import com.FemCoders.DropTheLink.video.Video;
 import jakarta.persistence.*;
 import lombok.*;
