+++ src/main/java/com/FemCoders/DropTheLink/user/Role.java
@@ -1,0 +1,5 @@
+package com.FemCoders.DropTheLink.user;
+
+public enum Role {
+    USER
+}
